requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiopg==1.3.1
aiosqlite==0.17.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
discord-components==2.1.1
discord.py==1.7.3
discordSuperUtils==0.2.2
idna==3.2
motor==2.5.1
multidict==5.1.0
Pillow==8.3.2
psycopg2-binary==2.9.1
py-cord @ git+https://github.com/Pycord-Development/pycord.git@cd70e292c59b6abec4f8f12a4e5889746d21705e
pycord-components==1.0.2
PycordPaginator==0.0.2
pycparser==2.20
pymongo==3.12.0
PyMySQL==0.9.3
PyNaCl==1.4.0
python-dotenv==0.19.0
pytz==2021.3
requests==2.26.0
six==1.16.0
spotipy==2.19.0
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
